MSPR_Pipeline/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
psycopg2-binary